PHPUnit: don't use __construct() / cross-version compat up to PHPUnit 7.x
The `__construct()` method in the PHPUnit `TestCase` class is used by that class itself.
Overloading it from a child class , while not problematic in PHPUnit 4.x, is problematic in later PHPUnit versions.

Better to use the `setUpBeforeClass()`or `setUp()` methods instead to set up fixtures.

With this change, the unit tests can now be run on PHPUnit 4 - 7.